refactor(audio): split decode io and prep
